requirements.txt
appnope==0.1.0
argh==0.26.2
asyncpg==0.18.2
asyncpool==1.0
backcall==0.1.0
decorator==4.3.0
distconfig==0.1.0
docopt==0.6.2
future==0.17.1
hbmqtt==0.9.5
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
marshmallow==2.16.3
parso==0.3.1
passlib==1.7.1
pathlib==1.0.1
pathtools==0.1.2
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
Pygments==2.3.1
pytoml==0.1.20
PyYAML==3.13
six==1.12.0
SQLAlchemy==1.2.15
traitlets==4.3.2
transitions==0.6.9
ujson==1.35
vyper-config==0.3.3
watchdog==0.9.0
wcwidth==0.1.7
websockets==7.0
